feat(cloudfront): Distribution - alternate domains (#9699)
Enable users to associate one or more alternate domain names with a
distribution. Originally, the plan was for this to come directly from the
certificate (if provided); however, the `ICertificate` does not include the
certificate domains.

Given adding a certificate without specifying the aliases is effectively a
no-op (the certificate will never be used), opted to make this an error instead
of a warning (a breaking change). Open to discussion and debate on this point.

BREAKING CHANGE: Distribution: `.domains` must be specified if `certificate` is provided.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*