feat(ec2): CloudFormation init for files, packages, sources, users, & groups (#9664)
Cloudformation Init (cfn-init) support was introduced in #9065 with a minimal
set of support for various init element types. This PR is a continuation of
that support (based on the original #8788), and adds the remaining init element
types: files, packages, sources, users, and groups.

With this PR, CloudFormation init support for EC2 instances is complete. A final
PR will be submitted to extend this support to auto-scaling groups (again, based
on the original work done in #8788).

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*