fix(cli): stack monitor reads complete stack history every 5 seconds
The stack monitor used to read all stack events on every tick (every 1-5
seconds).

That's right, every 5 seconds it would page through all events that had
ever happened to the stack, only to get to the last page to see if there
were new events. We were quadratic in our `DescribeStackEvents` API
call.

This behavior has probably contributed to our users getting throttled by
CloudFormation in the past, and only really shows pathological
behavior on stacks with a long history (which we ourselves don't
routinely manage using `cdk deploy`), which is why it had eluded
detection for so long.

The fix is to remember the token of the last page between calls,
so that we don't have to read through all pages just to get to the last
one.

Fixes #9470.