chore: cleanup awscli-compatible credential loading
There was a branch that obviously contained a bug, and yet all
tests worked fine.

It's because this code didn't do what I thought it did. The
`~/.aws/credentials` file is something we read *explicitly*, but the
`~/.aws/config` file is read *implicitly* by the
`SharedIniFileCredentials` if the right environment variable is set,
so we don't have to do that one explicitly.