fix(lambda-python): install rsync if necessary (#9763)
Fixes #9704.

`aws-lambda-python` originally used `rsync` to move lambda code into `/asset-output`, but I switched to `cp` in #9355 to resolve #9349 as `rsync` isn't installed in the python3.8 sam docker image.

This change introduces a Dockerfile derived from the `bundlingDockerImage` of the runtime which installs `rsync` if it's missing.

While this feels a bit heavy-handed, I had been planning to introduce a Dockerfile to setup a working pip cache after #9582 lands.

I'm also happy to take a different approach if you'd prefer, @eladb. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*