chore(cli): fix bugs introduced in stack monitor
In #9795, I was under the impression that the `DescribeStackEvents`
would list events in chronological order.

This was a misunderstanding resulting from not reading the documentation--
in fact, the events are returned in reverse chronological order.

The paging logic was therefore wrong and the events would be fed into
the activity printer in the wrong order, leading to incorrect progress
reporting.

This is a critical bug and must be fixed immediately!

While I was in there, fixed two other paper cuts around the
printer at the same time:

- Events occurring near the start of the monitor would be missed. This
  was because time-based filtering would be employed based on the
  client machine's local clock, which might be out of sync with the
  remote clock. Use the server's clock instead, by using the
  `CreationTime` of the ChangeSet as the start time.

- Events occurring near the end of the monitor would be missed. This
  was because `DescribeStackStatus` might return a completion state
  before the monitor had polled the events that led up to the completion.
  Fix by doing a final poll after the stop call has been received,
  so we are guaranteed to have seen all events.