fix(bootstrap): default to Admin permissions for same-account access
The current bootstrapping experience has sharp edges. It requires you
to pass `--cloudformation-execution-policies`. If you don't pass any,
it will successfully deploy a bootstrap stack that doesn't actually
work.

In effect, you are forced to constantly look up and copy/paste the ARN
for `AdministratorAccess` (because that's effectively what people
need to use anyway). We don't want to default to this policy for them,
because it equates to handing over the keys to your account to another
account, and you need to be fully aware that you're doing that.

In the case of bootstrapping an account just for use "by itself",
however (which is the 90% use case for bootstrapping),
the trust boundary is very clear and there's no risk of privilege
escalation between accounts (there might still be risk of privilege
escalation between different people using different IAM users in
the same AWS account, but that is not a recommended setup).

Add an optimization where in the case of a "simple", non-cross account
bootstrap, we'll default to the AdministratorAccess ARN for you.

Once you establish `--trust` with another account, we'll still
force you to spell out the execution policy you'll want to use though.

Fixes #8571.

Also in this PR:

- The `blockPublicAccessConfiguration` would be reset on every
  re-bootstrap, instead of reusing the previous setting.
- The `enableTerminationProtection` would be reset on every
  re-bootstrap, instead of reusing the previous setting.