chore(s3-deployment): update README to delineate the CloudFront options (#9893)
The existing `CloudFrontWebDistribution` construct needs to be configured
quite differently depending on whether the backing bucket is configured for
website hosting or not; this can lead to confusion and incorrect results if
the wrong origin type is used.

This doc update explicitly calls out the different options, including the newer
experimental construct that hides this complexity entirely.

fixes #7434


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*