chore(ec2): remove unneeded override LookedUpVpc.vpnGatewayId
The `vpnGatewayId` property is already declared as a dynamic property
(a getter) in `VpcBase`, and it is being overridden by a "plain"
property declaration. This pattern is disallowed in TypeScript 4 due to
the risk of runtime errors when this is done incorrectly (see more
details in microsoft/TypeScript#37894).

In any case, the override was likely a mistake, as no value was ever set
to the overridden property (values were correctly set on
`this._vpnGatewayId`, which backs the dynamic getter on `VpcBase`).