chore: convert ELB and ELBv2 tests from nodeunit to jest (#9967)
Part of my continuing campaign to update everything to Jest. There were a small
enough number of tests to opt for full rewrites, rather than using the
`nodeunitshim` helper.

One little hack for the diff's sake. I added a dummy `describe('tests')` group
to the top of each test to keep the indentation the same as with nodeunit. This
fixed git's delta detection algorithm to only show actual changes, rather than
showing the old and new files as completely unrelated.