fix(lambda-nodejs): incorrect working directory for local bundling (#9870)
Execute local bundling from the directory containing the entry file.

Without this change, in a monorepo with multiple `package.json` files
or when consuming a module exposing a construct, Parcel doesn't look for
the right `package.json`.

Also fix a regression introduced in #9632 for the working directory in
the container.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*