fix(lambda): grantInvoke fails on second invocation (#9960)
If `grantInvoke()` is called twice for the same principal, the second call fails
due to attempting to create two `CfnPermission` nodes with the same id. This
(simple) fix skips the second creation if the node already exists.

A more robust check would be to check the existing `CfnPermission`, comparing
every field, skipping creation if the two are identical and throwing an error
otherwise, as well as handling that in the upstream `grantInvoke` call. I opted
for the simpler solution for now, but willing to take arguments for something
more complex.

I also nested the existing grantInvoke tests for future readability. The tests weren't changed,
just the last one added.

fixes #8553


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*