feat(cloudfront): import existing CloudFrontWebDistributions (#10007)
Existing distributions can already been imported with the new (experimental)
`Distribution` construct, but the functionality had never been backported to
the stable construct. This makes it slightly more discoverable for users who
are using the `CloudFrontWebDistribution` construct.

Note: Opted in the README to only update the Distribution section, as it felt
redundant to have the same README section twice for the two different classes.

fixes #5607


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*