chore(monocdk-experiment): `constructs` should NOT be a dependency
Move `constructs` from `dependencies` to `devDependencies`.

The whole point of the monocdk reorganization is that dependencies whose
types are exposed in a package's public API (of which `constructs`,
delivering the `Construct` class, is definitely one), should NOT be
in `dependencies` but in `peerDependencies`.

It was probably added to `dependencies` to make the build work. The
correct place to put a dependency for build purposes is in
`devDependencies.`