fix(aws-stepfunctions-tasks): missing permission to get build status (#10081)
The execution role for the CodeBuild StartBuild task is missing permissions to check the status of the running build. It results in a timeout of the step functions.

Adding `codebuild:BatchGetBuilds` and `codebuild:BatchGetReports` following this [AWS news blog article](https://aws.amazon.com/blogs/aws/new-building-a-continuous-integration-workflow-with-step-functions-and-aws-codebuild/) solves the issue.

Closes #8043

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*