chore(core,cx-api): centralize defaults for future flags
This is in preparation for v2.0 in which future flags will be turned on
by default.

For a period of few months, two active branches `master` and `v2-master`
will be present with continuous merges from former to latter. This
change will reduce the number of merge conflicts between the branches
after the defaults have been flipped.