chore(core,cx-api): centralize defaults for future flags (#10124)
This is in preparation for v2.0 in which future flags will be turned on
by default.

For a period of few months, two active branches `master` and `v2-master`
will be present with continuous merges from former to latter. This
change will reduce the number of merge conflicts between the branches
after the defaults have been flipped.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*