fix(cli): metadata not recorded in templates >50k
The CDK Metadata Resource is added by the CLI, and only added
to the in-memory representation of the template.

This in-memory representation is only used in the `CreateChangeSet`
call if the full template body is <=50kB.

If the template is larger than this size, the template file that's on
disk will be uploaded to the bootstrap S3 bucket and referenced.
However, that template will not have had the metadata resource
injected and so the stack will not be registered.

This PR:

- Provisionally fixes this issue by flushing the in-memory modifications
  back out to disk.
- Also adds the metadata resources to stacks found in nested Cloud
  Assemblies, such as the stacks deployed via CDK Pipelines.

There is some jank around new-style synthesis stacks: the protocol
specifies that those stacks have synthesized their templates as assets,
and those will not (necessarily) be processed by the CLI, although
it might coincidentally work if the files are the same.

The proper fix to this whole mess is to have the framework inject the
metadata resource at synthesis time. That fix is forthcoming, but this
current one is a good stopgap until that time.