fix(codepipeline): cross-region support stack requires bootstrapping
The cross-region support stack was intended to not require
bootstrapping. The recent "version check" feature added to the
default synthesizer did lead to bootstrapping being required (because
an SSM parameter needed to be provisioned in the target environment).

Make adding that check switchable and switch it off for the support
stack (which uses a `BootstraplessSynthesizer`).

Fixes #10215.