fix(lambda-nodejs): permission denied, mkdir '/.parcel-cache' (#10181)
When the project root (=`/asset-input` in the container) doesn't contain
a lock file (npm or yarn), Parcel thinks that `/asset-input` is part of
a monorepo with the root at `/` because the image installs Parcel at
`/`. It will then try to create the cache at `/`.

Ensure no lock file is created during installation of Parcel in the
Docker image.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*