chore: aws-cdk-lib as a private package (#10161)
Create a private package `aws-cdk-lib` as a replica of
`monocdk-experiment`. This private package will not be published.

The build logic in `monocdk-experiement` is refactored into a new build
tool - `ubergen` - and is shared between `monocdk-experiment` and the
new `aws-cdk-lib`.

Motivation
As we get closer to launch v2, there is going to be a period of several months
where there will be two active branches, one tracking v1 and the other tracking
v2. 

The main motivation for this change is merge conflict reduction during this time.

During this period, we're going to work on polishing up the experience of
`aws-cdk-lib` and may build more tests and monitoring. Without this change,
the branch tracking v1 will have several places referring to this package as
`monocdk-experiment` while the branch tracking v2 will refer to this as
`aws-cdk-lib` leading to increased merge conflicts.

Once this change is checked in, the only difference between the two branches
should be the `private` field in the `package.json` file. This should significantly
reduce conflicts between the two branches.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*