chore(cli): mark `cdk deploy --tags` for removal in v2 (#10126)
`cdk deploy --tags` does a one-time override of tags found in the
Cloud Assembly. The tags will be reset on the next deployment.

Specifying tags at deploy time is at odds with the desired-state
configuration that the CDK promotes, and the 2 sources of tags
don't work together well.

We should stop supporting this.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*