fix(core): DefaultSynthesizer breaks this.node.setContext() on Stack
Since the `DefaultSynthesizer` recently started adding constructs into
the Stack at construction time, that broke `this.node.setContext()`
which requires that no children have been added to a construct yet
when its context is being modified.

To fix this, add the constructs just-in-time just before the stack
is being synthesized.

In order to give the `DefaultStackSynthesizer` a chance to modify
the stack's construct tree before its template is being written out,
the Synthesizer is now in full control of the order in which things
happen.

Change the call tree from:

```
synthesizeTree
  - stack._synthesizeTemplate
      - (write template)
      - this.synthesizer.synthesizeStackArtifacts
          - (register artifacts)
```

To:

```
synthesizeTree
  - stack.synthesizer.synthesize
      - stack._synthesizeTemplate
          - (write template)
      - (register artifacts)
```

All APIs involved in this call tree are either `@experimental` or
`@internal`.

BREAKING CHANGE: custom implementations of `IStackSynthesizer`
must now implement `synthesize()` instead of
`synthesizeStackArtifacts()`.