fix(codepipeline): cross-region support stack requires bootstrapping (#10217)
The cross-region support stack was intended to not require
bootstrapping. The recent "version check" feature added to the
default synthesizer did lead to bootstrapping being required (because
an SSM parameter needed to be provisioned in the target environment).

Make adding that check switchable and switch it off for the support
stack (which uses a `BootstraplessSynthesizer`).

Fixes #10215.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*