fix(cloudfront): comment for origin access identity is too long
The OAI gets a comment that includes the name of the origin it is attached to.
With longer distribution and origin names, this comment can overflow the
(undocumented) 128-character limit. This fix truncates the comment.

I also updated `oai.test.ts` from using the `nodeunitshim` to actually be
written as Jest tests.

fixes #10211