fix(kms): do not change the principal to root for imported resources in dependent Stacks (#10299)
We have logic in KMS Key that checks whether the grantee is from a Stack that depends on the Key's Stack.
It's required because KMS validated that the principals contained in its Key policy actually exist,
and fails if they don't, so in that case, we switch to using the root principal instead.
However, that logic only makes sense for newly created resources;
for imported resources, like those with `Role.fromRoleArn`,
they already exist, so no need to make this switch.

Fixes  #10166.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*